requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.73.0
h11==0.13.0
idna==3.3
iso8601==0.1.16
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pypika-tortoise==0.1.3
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tortoise-orm==0.18.1
typing_extensions==4.0.1
uvicorn==0.17.0.post1

pytest~=7.0.0
httpx~=0.22.0